seed rand with processor clock
